refactored model managers into their own file
