+++ plan/views.py
@@ -1,10 +1,7 @@
-from django.db.models import Subquery, OuterRef, Avg
-
 from courses.views import CourseList
 from .search import TypedSearchBackend
 
 from courses.models import Requirement
-from review.models import ReviewBit
 from .serializers import CourseListSearchSerializer
 
 
@@ -24,20 +21,5 @@ class CourseListSearch(CourseList):
             except Requirement.DoesNotExist:
                 return queryset.none()
             queryset = queryset.filter(id__in=requirement.satisfying_courses.all())
-
-        fields = ['course_quality', 'difficulty', 'instructor_quality']
-
-        queryset = queryset.annotate(**{
-            field: Subquery(
-                ReviewBit.objects.filter(review__section__course__full_code=OuterRef('full_code'),
-                                         field=field)
-                .values('field')
-                .order_by()
-                .annotate(avg=Avg('score'))
-                .values('avg')[:1]
-
-            )
-            for field in fields
-        })
 
         return queryset
